Minor updates to imgui lib
